refactor(objectives): move PaperSkim contracts into discovery
Why:
- PaperSkim window extraction and signal reconciliation each required reading generic prompt, schema, and extractor monoliths to understand one scientific judgment.

What:
- Co-locate each judgment's prompt, response contract, validation, repair, bounds, and model call under discovery while retaining one shared structured-response client.

Impact:
- Source batching, retries, consolidation, prompt content, model budgets, scientific outputs, persistence, and public contracts remain unchanged.

Verification:
- All 152 PaperSkim contract tests and 12 Objective workflow and collection-build integration tests passed.
