+++ backend/application/core/objectives/discovery/__init__.py
@@ -1,0 +1,1 @@
+"""Research Objective discovery responsibilities."""
