+++ backend/application/core/objectives/prompts.py
@@ -3,8 +3,6 @@ from __future__ import annotations
 import json
 from typing import Any
 
-PAPER_SKIM_PROMPT_VERSION = "paper_skim.v2"
-PAPER_SIGNAL_RECONCILIATION_PROMPT_VERSION = "paper_signal_reconciliation.v3"
 RESEARCH_AXIS_CANONICALIZATION_PROMPT_VERSION = "research_axis_canonicalization.v1"
 OBJECTIVE_PAPER_FRAME_PROMPT_VERSION = "objective_paper_frame.v2"
 OBJECTIVE_EVIDENCE_ROUTE_PROMPT_VERSION = "objective_evidence_route.v1"
@@ -21,32 +19,6 @@ Non-negotiable rules:
 - Do not infer material systems from filenames.
 - Prefer fewer, higher-signal outputs over speculative coverage.
 - Research objectives must be question-shaped. Do not return a plain material list.
-""".strip()
-
-
-_PAPER_SKIM_SYSTEM_PROMPT = """
-You are screening one bounded Source window for a traceable literature map.
-
-Non-negotiable rules:
-- This is high-recall study-design screening, not final fact extraction or synthesis.
-- Return exactly one JSON object and nothing else.
-- Scientific labels must be supported by supplied Source-unit content.
-- Copy only supplied `source_unit_id` values; never invent or rewrite an id.
-- Do not infer material systems from filenames or section names.
-""".strip()
-
-
-_PAPER_SIGNAL_RECONCILIATION_SYSTEM_PROMPT = """
-You are adjudicating one bounded candidate neighborhood within one paper.
-
-Non-negotiable rules:
-- This batch has exactly one outcome anchor and candidate variable signals selected by the backend.
-- This is paper-level membership adjudication, not cross-paper grouping or final synthesis.
-- Return exactly one JSON object and nothing else.
-- Link signals only when their supplied excerpts and contexts support one study design.
-- Copy only supplied `signal_id` values; never invent scientific labels or ids.
-- Preserve ambiguity by returning an unresolved signal instead of guessing a link.
-- Account only for the current batch; the backend derives final whole-paper accounting.
 """.strip()
 
 
@@ -263,172 +235,6 @@ Return exactly `{"findings":[]}` or one object shaped as
 `"context_evidence_ids":[],"mechanisms":[]}]}`. Use empty arrays when
 annotations are absent and no extra keys.
 """.strip()
-
-
-def build_paper_skim_prompt(payload: dict[str, Any]) -> tuple[str, str]:
-    user_prompt = (
-        "TASK MODEL\n"
-        "Extract source-supported paper studies from one bounded Source window. "
-        "This is high-recall study-structure extraction, not objective wording, "
-        "collection grouping, final measurement extraction, or synthesis.\n\n"
-        "INPUT SCHEMA\n"
-        "- `document_id` and `title` identify the Source paper.\n"
-        "- `window_id` is this bounded window's identity; `window_role` is one of "
-        "overview, methods, results, conclusion, or unknown.\n"
-        "- `source_units` contains every Source item assigned to this window. Each "
-        "unit has an opaque `source_unit_id`, stable Source kind/reference, section "
-        "path, and text or caption content. Source identity is provenance; content "
-        "is the scientific authority.\n"
-        "- `document_profile` is a coarse paper-level classification hint.\n"
-        "This is one incomplete view of the paper; absence from this window is not "
-        "evidence of absence elsewhere. Window metadata describes input provenance "
-        "and must not appear in output.\n\n"
-        f"Input JSON:\n{json.dumps(payload, ensure_ascii=False, indent=2)}\n\n"
-        "DECISION PROCESS\n"
-        "1. Classify the paper role from explicit study-design signals.\n"
-        "2. Identify each distinct experiment, observation, or model represented in "
-        "this window. Keep different specimens, tests, processes, comparators, or "
-        "experiment labels as separate studies.\n"
-        "3. For each study, record design_type and claim_scope. Only claims about the "
-        "paper's own work use claim_scope=current_work; review synthesis and cited "
-        "background remain synthesis or background.\n"
-        "4. Express every factor and outcome as a neutral scientific axis. A factor "
-        "names what was varied, compared, or modeled, not its tested levels. An "
-        "outcome names what was measured or predicted, not the result direction, "
-        "value, or comparison sentence.\n"
-        "5. Within each study, return one relationship per outcome. `varied_factors` "
-        "must contain the full jointly varied, compared, or modeled factor set. Never "
-        "split a joint-factor experiment into isolated effects.\n"
-        "6. Record material, process, sample, test, comparator, and fixed-condition "
-        "context only when explicitly supported.\n"
-        "7. Copy every unique Source-unit id that directly supports each relationship "
-        "or unresolved signal. Each item may contain at most 12 unique "
-        "`source_unit_ids`.\n"
-        "8. When the window explicitly identifies a varied/modeled variable but no "
-        "response, or a measured/predicted outcome but no changed variable, return "
-        "the explicit axis in `unresolved_signals` for paper-level reconciliation.\n"
-        "9. Use evidence density, confidence, and warnings to expose incomplete or "
-        "ambiguous input rather than filling gaps.\n\n"
-        "HARD RULES\n"
-        "- Return only the schema object. Return every distinct, explicitly supported "
-        "study and relationship visible in this Source window; do not discard one "
-        "because another appears more central.\n"
-        "- Extract only relationships supported inside this window. Do not guess what "
-        "another section may contain. Repeating a study fragment found in another "
-        "window is acceptable; backend consolidation is authoritative.\n"
-        "- Never move a factor, outcome, or context between studies.\n"
-        "- Every relationship and unresolved signal must copy `source_unit_ids` that "
-        "directly support it. Do not return an id absent from `source_units`, repeat an "
-        "id inside one item, or return more than 12 IDs for one item.\n"
-        "- Do not repeat an axis in `unresolved_signals` when it is already part of a "
-        "complete relationship in this window. Material and fixed process context are "
-        "not partial variable/outcome signals.\n"
-        "- Do not generate a research question or collection-level objective.\n"
-        "- Do not infer scientific content from filenames or generic section names.\n"
-        "- Return empty arrays rather than guessing unsupported study structure.\n\n"
-        "OUTPUT CONTRACT\n"
-        "- Return `studies`, `unresolved_signals`, doc_role, evidence_density, "
-        "confidence, warnings, and `output_saturated`. A study has "
-        "experiment/design/context fields and "
-        "one or more relationships. A relationship has `varied_factors`, one "
-        "`outcome`, `source_unit_ids`, and confidence.\n"
-        "- Return up to 8 studies, up to 8 relationships per study, and up to 12 "
-        "unresolved signals. If every visible fact fits, set "
-        "`output_saturated=false`. If any distinct supported study, relationship, "
-        "or signal would exceed those limits, set `output_saturated=true`; the "
-        "backend will split and retry the Source window. Never silently choose a "
-        "subset.\n"
-        "- Each relationship and unresolved signal returns at most 12 unique "
-        "`source_unit_ids`, matching the maximum Source units in one input window.\n"
-        "- Return up to 2 `warnings`, each at most 240 characters.\n"
-        "- Keep each value concise and preserve exact joint-factor-to-outcome links.\n\n"
-        "BOUNDARY EXAMPLES\n"
-        "- Supported relationship: text says laser power was varied and relative "
-        "density and porosity were measured. Return one study with two relationships; "
-        "each has varied_factors=['laser power'] and one distinct outcome.\n"
-        "- Joint factors: power and speed changed together. Keep "
-        "varied_factors=['power','speed']; do not emit isolated power or speed effects.\n"
-        "- Factor levels: specimens use L-VED, M-VED, and H-VED. Return "
-        "varied_factors=['volumetric energy density']; keep the level names in Source "
-        "evidence rather than returning them as three factors.\n"
-        "- Result clause: text says fatigue strength decreases with lower VED. Return "
-        "outcome='fatigue strength'; the decrease and condition belong to later "
-        "Evidence extraction, not the outcome axis.\n"
-        "- Incomplete relationship: a Methods window names laser power but no "
-        "measured or predicted response. Return `studies=[]`; do not "
-        "borrow an outcome from another section. Return the explicit axis in "
-        "`unresolved_signals` with its supporting Source-unit id.\n"
-        "- No study signal: a unit contains only general background. Return no study "
-        "or unresolved signal for that unit.\n"
-        "- Separate relationships: one experiment links scan speed to porosity and "
-        "another links heat treatment to yield strength. Return two studies."
-    )
-    return _PAPER_SKIM_SYSTEM_PROMPT, user_prompt
-
-
-def build_paper_signal_reconciliation_prompt(
-    payload: dict[str, Any],
-) -> tuple[str, str]:
-    user_prompt = (
-        "TASK MODEL\n"
-        "Decide whether the candidate variables in one bounded candidate neighborhood "
-        "belong to the same experiment or model as its exactly one outcome anchor. "
-        "This is membership adjudication, not scientific-field generation or "
-        "whole-paper discovery.\n\n"
-        "INPUT SCHEMA\n"
-        "- `document_id` identifies the one paper.\n"
-        "- `signals` contains exactly one outcome anchor and one or more candidate "
-        "variables. Each has a backend-owned `signal_id`, exact label, known scientific "
-        "context, and bounded Source excerpts with stable Source-unit positions.\n"
-        "- Signals omitted from this request are outside the current batch; omitted "
-        "paper signals are outside this batch, not negative evidence.\n"
-        "- Source excerpts are the only authority for deciding whether signals belong "
-        "to the same experiment or model.\n\n"
-        f"Input JSON:\n{json.dumps(payload, ensure_ascii=False, separators=(',', ':'))}\n\n"
-        "DECISION PROCESS\n"
-        "1. Identify the single outcome anchor and evaluate each candidate variable "
-        "against it.\n"
-        "2. Compare their material, process, condition, experiment, sample, and "
-        "section evidence. Do not link signals merely because they occur in the same paper.\n"
-        "3. Form a relationship only when the excerpts support the variable as varied, "
-        "compared, or modeled and the outcome as the response for that same study design.\n"
-        "4. Return at most one study for this neighborhood. Relationships may share the "
-        "outcome anchor only when the Source explicitly supports that membership.\n"
-        "5. Do not reason about omitted signals or attempt to reconstruct the whole paper.\n"
-        "6. Include a rejected candidate once in `unresolved_signals` when a concise "
-        "scientific reason is visible. The backend treats every omitted input signal "
-        "as unresolved, so never invent a reason merely to repeat an ID.\n\n"
-        "HARD RULES\n"
-        "- In every relationship, copy only input `signal_id` values and include at "
-        "least one variable signal and one outcome signal.\n"
-        "- Keep `signal_ids` unique inside each relationship, and never return the "
-        "same signal membership more than once. Relationship membership is unordered; "
-        "reversing the same IDs does not create another relationship.\n"
-        "- Never combine incompatible materials, processes, samples, tests, or "
-        "experiments. Ambiguous proximity is not a link.\n"
-        "- Do not output labels, contexts, Source locators, questions, or new scientific "
-        "fields; the backend derives them from selected signals.\n"
-        "- Do not mark a signal unresolved if it appears in a relationship. Backend "
-        "relationship acceptance is authoritative when the response repeats an ID.\n\n"
-        "BOUNDARY EXAMPLES\n"
-        "- Methods variable and Results outcome: Methods says laser power was varied "
-        "for LPBF 316L specimens; Results says relative density was measured for those "
-        "power conditions. Link both signal ids into one relationship.\n"
-        "- Different experiments: Methods describes heat-treatment temperature for "
-        "tensile coupons, while Results reports corrosion potential for as-built "
-        "electrochemical specimens. Return both signals unresolved; do not link them.\n"
-        "- Ambiguous result: Results lists hardness without identifying which of two "
-        "independent process studies it belongs to. Keep the hardness signal unresolved.\n\n"
-        "- Duplicate membership: [variable-a,outcome-a] and "
-        "[outcome-a,variable-a] are the same relationship. Return it once.\n\n"
-        "OUTPUT CONTRACT\n"
-        "Return exactly `studies` and `unresolved_signals`. Return at most one study, "
-        "up to 11 relationships, and up to 12 unresolved signals. Each relationship "
-        "contains only `signal_ids` and `confidence`; each unresolved item has one "
-        "`signal_id` and a concise `reason`. The backend derives final whole-paper "
-        "accounting after all candidate batches finish. Return empty arrays when appropriate."
-    )
-    return _PAPER_SIGNAL_RECONCILIATION_SYSTEM_PROMPT, user_prompt
 
 
 def build_research_axis_canonicalization_prompt(
