+++ backend/application/core/objectives/schemas.py
@@ -13,25 +13,12 @@ from pydantic import (
     model_validator,
 )
 
-_PAPER_SKIM_DOC_ROLES = {
-    "experimental",
-    "review",
-    "modeling",
-    "mixed",
-    "uncertain",
-}
-_PAPER_SKIM_EVIDENCE_DENSITIES = {"high", "medium", "low", "unknown"}
 PAPER_OBJECTIVE_CANDIDATE_OUTPUT_LIMITS = {
     "material_scope": (8, 80),
     "process_context": (4, 80),
     "variables": (4, 80),
     "outcomes": (8, 80),
 }
-PAPER_SKIM_WARNING_LIMIT = (2, 240)
-PAPER_SKIM_STUDY_LIMIT = 8
-PAPER_SKIM_RELATIONSHIP_LIMIT = 8
-PAPER_SKIM_UNRESOLVED_SIGNAL_LIMIT = 12
-PAPER_SKIM_SOURCE_UNIT_LIMIT = 12
 _OBJECTIVE_FRAME_RELEVANCE = {"high", "medium", "low", "irrelevant", "uncertain"}
 _OBJECTIVE_FRAME_PAPER_ROLES = {
     "primary_experiment",
@@ -125,394 +112,6 @@ class _StrictModel(BaseModel):
         return cls.model_fields["epistemic_status"].get_default(
             call_default_factory=True
         )
-
-
-class StructuredPaperStudyRelationship(_StrictModel):
-    varied_factors: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(min_length=1, max_length=8)
-    outcome: Annotated[str, Field(min_length=1, max_length=80)]
-    source_unit_ids: list[
-        Annotated[str, Field(min_length=1, max_length=160)]
-    ] = Field(
-        min_length=1,
-        max_length=PAPER_SKIM_SOURCE_UNIT_LIMIT,
-    )
-    confidence: float = 0.0
-
-    @field_validator("varied_factors", "source_unit_ids", mode="before")
-    @classmethod
-    def _normalize_lists(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-    @model_validator(mode="after")
-    def _validate_source_unit_ids(self) -> "StructuredPaperStudyRelationship":
-        normalized = [value.strip() for value in self.source_unit_ids]
-        if any(not value for value in normalized):
-            raise ValueError("paper relationship source-unit ids cannot be empty")
-        if len(normalized) != len(set(normalized)):
-            raise ValueError("paper relationship source-unit ids must be unique")
-        return self
-
-
-class StructuredPaperStudy(_StrictModel):
-    experiment_label: str | None = Field(default=None, max_length=120)
-    design_type: Literal[
-        "experimental",
-        "observational",
-        "modeling",
-        "mixed",
-        "uncertain",
-    ] = "uncertain"
-    claim_scope: Literal[
-        "current_work",
-        "synthesis",
-        "background",
-        "uncertain",
-    ] = "uncertain"
-    material_scope: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=8)
-    process_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    sample_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    test_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    comparator: str | None = Field(default=None, max_length=160)
-    fixed_conditions: list[
-        Annotated[str, Field(max_length=120)]
-    ] = Field(default_factory=list, max_length=12)
-    relationships: list[StructuredPaperStudyRelationship] = Field(
-        min_length=1,
-        max_length=PAPER_SKIM_RELATIONSHIP_LIMIT,
-    )
-    confidence: float = 0.0
-
-    @field_validator(
-        "material_scope",
-        "process_context",
-        "sample_context",
-        "test_context",
-        "fixed_conditions",
-        "relationships",
-        mode="before",
-    )
-    @classmethod
-    def _normalize_lists(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-    def identity_key(
-        self,
-        source_keys: Mapping[str, tuple[str, str]] | None = None,
-    ) -> tuple[object, ...]:
-        def normalized_values(values: list[str]) -> tuple[str, ...]:
-            return tuple(
-                sorted(
-                    {
-                        str(value).strip().casefold()
-                        for value in values
-                        if str(value).strip()
-                    }
-                )
-            )
-
-        relationships = tuple(
-            sorted(
-                (
-                    normalized_values(relationship.varied_factors),
-                    relationship.outcome.strip().casefold(),
-                    tuple(
-                        sorted(
-                            {
-                                source_keys.get(
-                                    source_unit_id.strip(),
-                                    ("source_unit", source_unit_id.strip()),
-                                )
-                                if source_keys is not None
-                                else ("source_unit", source_unit_id.strip())
-                                for source_unit_id in relationship.source_unit_ids
-                                if source_unit_id.strip()
-                            }
-                        )
-                    ),
-                )
-                for relationship in self.relationships
-            )
-        )
-        return (
-            self.design_type,
-            self.claim_scope,
-            self.experiment_label.strip().casefold()
-            if self.experiment_label
-            else None,
-            normalized_values(self.material_scope),
-            normalized_values(self.process_context),
-            normalized_values(self.sample_context),
-            normalized_values(self.test_context),
-            self.comparator.strip().casefold() if self.comparator else None,
-            normalized_values(self.fixed_conditions),
-            relationships,
-        )
-
-
-class StructuredPaperStudySignal(_StrictModel):
-    signal_type: Literal["variable", "outcome"]
-    label: Annotated[str, Field(min_length=1, max_length=80)]
-    experiment_label: str | None = Field(default=None, max_length=120)
-    design_type: Literal[
-        "experimental",
-        "observational",
-        "modeling",
-        "mixed",
-        "uncertain",
-    ] = "uncertain"
-    claim_scope: Literal[
-        "current_work",
-        "synthesis",
-        "background",
-        "uncertain",
-    ] = "uncertain"
-    material_scope: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=8)
-    process_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    sample_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    test_context: list[
-        Annotated[str, Field(max_length=80)]
-    ] = Field(default_factory=list, max_length=4)
-    comparator: str | None = Field(default=None, max_length=160)
-    fixed_conditions: list[
-        Annotated[str, Field(max_length=120)]
-    ] = Field(default_factory=list, max_length=12)
-    source_unit_ids: list[
-        Annotated[str, Field(min_length=1, max_length=160)]
-    ] = Field(
-        min_length=1,
-        max_length=PAPER_SKIM_SOURCE_UNIT_LIMIT,
-    )
-    confidence: float = 0.0
-
-    @field_validator(
-        "material_scope",
-        "process_context",
-        "sample_context",
-        "test_context",
-        "fixed_conditions",
-        "source_unit_ids",
-        mode="before",
-    )
-    @classmethod
-    def _normalize_lists(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-    @model_validator(mode="after")
-    def _validate_source_unit_ids(self) -> "StructuredPaperStudySignal":
-        normalized = [value.strip() for value in self.source_unit_ids]
-        if any(not value for value in normalized):
-            raise ValueError("paper signal source-unit ids cannot be empty")
-        if len(normalized) != len(set(normalized)):
-            raise ValueError("paper signal source-unit ids must be unique")
-        return self
-
-
-class StructuredPaperSkim(_StrictModel):
-    doc_role: Literal["experimental", "review", "modeling", "mixed", "uncertain"] = (
-        "uncertain"
-    )
-    studies: list[StructuredPaperStudy] = Field(
-        default_factory=list,
-        max_length=PAPER_SKIM_STUDY_LIMIT,
-    )
-    unresolved_signals: list[StructuredPaperStudySignal] = Field(
-        default_factory=list,
-        max_length=PAPER_SKIM_UNRESOLVED_SIGNAL_LIMIT,
-    )
-    output_saturated: bool = False
-    evidence_density: Literal["high", "medium", "low", "unknown"] = "unknown"
-    confidence: float = 0.0
-    warnings: list[
-        Annotated[str, Field(max_length=PAPER_SKIM_WARNING_LIMIT[1])]
-    ] = Field(
-        default_factory=list,
-        max_length=PAPER_SKIM_WARNING_LIMIT[0],
-    )
-
-    @model_validator(mode="before")
-    @classmethod
-    def _downgrade_relationships_without_factors(cls, value: object) -> object:
-        if not isinstance(value, Mapping):
-            return value
-        studies = value.get("studies")
-        unresolved_signals = value.get("unresolved_signals")
-        if not isinstance(studies, list):
-            return value
-        if unresolved_signals is not None and not isinstance(
-            unresolved_signals, list
-        ):
-            return value
-
-        retained_studies: list[object] = []
-        downgraded_signals: list[dict[str, object]] = []
-        changed = False
-        for study in studies:
-            if not isinstance(study, Mapping):
-                retained_studies.append(study)
-                continue
-            relationships = study.get("relationships")
-            if not isinstance(relationships, list):
-                retained_studies.append(study)
-                continue
-
-            retained_relationships: list[object] = []
-            study_changed = False
-            for relationship in relationships:
-                if not isinstance(relationship, Mapping):
-                    retained_relationships.append(relationship)
-                    continue
-                varied_factors = relationship.get("varied_factors")
-                if not isinstance(varied_factors, list) or any(
-                    str(item).strip() for item in varied_factors
-                ):
-                    retained_relationships.append(relationship)
-                    continue
-                outcome = str(relationship.get("outcome") or "").strip()
-                source_unit_ids = relationship.get("source_unit_ids")
-                if (
-                    not outcome
-                    or not isinstance(source_unit_ids, list)
-                    or not any(str(item).strip() for item in source_unit_ids)
-                ):
-                    retained_relationships.append(relationship)
-                    continue
-
-                signal = {
-                    "signal_type": "outcome",
-                    "label": outcome,
-                    "source_unit_ids": list(source_unit_ids),
-                    "confidence": relationship.get(
-                        "confidence", study.get("confidence")
-                    ),
-                }
-                for field_name in (
-                    "experiment_label",
-                    "design_type",
-                    "claim_scope",
-                    "material_scope",
-                    "process_context",
-                    "sample_context",
-                    "test_context",
-                    "comparator",
-                    "fixed_conditions",
-                ):
-                    if field_name in study:
-                        signal[field_name] = study[field_name]
-                downgraded_signals.append(signal)
-                study_changed = True
-                changed = True
-
-            if retained_relationships or not study_changed:
-                retained_study = dict(study)
-                retained_study["relationships"] = retained_relationships
-                retained_studies.append(retained_study)
-
-        if not changed:
-            return value
-        normalized = dict(value)
-        normalized["studies"] = retained_studies
-        normalized["unresolved_signals"] = [
-            *(unresolved_signals or []),
-            *downgraded_signals,
-        ]
-        return normalized
-
-    @field_validator(
-        "studies",
-        "unresolved_signals",
-        "warnings",
-        mode="before",
-    )
-    @classmethod
-    def _normalize_lists(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-    @field_validator("doc_role", mode="before")
-    @classmethod
-    def _normalize_doc_role(cls, value: object) -> str:
-        return _normalize_underscored_choice(
-            value,
-            allowed=_PAPER_SKIM_DOC_ROLES,
-            default="uncertain",
-        )
-
-    @field_validator("evidence_density", mode="before")
-    @classmethod
-    def _normalize_evidence_density(cls, value: object) -> str:
-        return _normalize_underscored_choice(
-            value,
-            allowed=_PAPER_SKIM_EVIDENCE_DENSITIES,
-            default="unknown",
-        )
-
-    @model_validator(mode="after")
-    def _validate_study_identities(self) -> "StructuredPaperSkim":
-        study_identities = [study.identity_key() for study in self.studies]
-        if len(study_identities) != len(set(study_identities)):
-            raise ValueError("studies contain duplicate study identities")
-        return self
-
-
-class StructuredPaperSignalRelationship(_StrictModel):
-    signal_ids: list[Annotated[str, Field(max_length=80)]] = Field(
-        min_length=2,
-        max_length=12,
-    )
-    confidence: float = 0.0
-
-    @field_validator("signal_ids", mode="before")
-    @classmethod
-    def _normalize_signal_ids(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-
-class StructuredUnresolvedPaperSignal(_StrictModel):
-    signal_id: Annotated[str, Field(min_length=1, max_length=80)]
-    reason: Annotated[str, Field(min_length=1, max_length=240)]
-
-
-class StructuredPaperSignalStudy(_StrictModel):
-    relationships: list[StructuredPaperSignalRelationship] = Field(
-        min_length=1,
-        max_length=11,
-    )
-
-    @field_validator("relationships", mode="before")
-    @classmethod
-    def _normalize_relationships(cls, value: object) -> object:
-        return _normalize_list_container(value)
-
-
-class StructuredPaperSignalReconciliation(_StrictModel):
-    studies: list[StructuredPaperSignalStudy] = Field(
-        default_factory=list,
-        max_length=1,
-    )
-    unresolved_signals: list[StructuredUnresolvedPaperSignal] = Field(
-        default_factory=list,
-        max_length=12,
-    )
-
-    @field_validator("studies", "unresolved_signals", mode="before")
-    @classmethod
-    def _normalize_lists(cls, value: object) -> object:
-        return _normalize_list_container(value)
 
 
 class StructuredAxisPairDecision(_StrictModel):
